+++ src/integrations/FreeformIntegration.php
@@ -13,25 +13,28 @@ class FreeformIntegration {
 
   public function parse(string $integration): void {
     $this->integration = $integration;
-    Event::on(\Solspace\Freeform\Form\Form::class, \Solspace\Freeform\Form\Form::EVENT_SUBMIT, function (\Solspace\Freeform\Events\Forms\SubmitEvent $e){      
-      $params = [
-        'content' => []
-      ];
-      foreach($e->getForm()->getFields() as $field){
-        switch(get_class($field)){
-          case 'Solspace\Freeform\Fields\Implementations\EmailField':
-            $params['email'] = $field->getValue();
-            break;
-          case 'Solspace\Freeform\Fields\Implementations\TextareaField':
-            $params['content'][] = $field->getValue();
-            break;
+    Event::on(\Solspace\Freeform\Form\Form::class, \Solspace\Freeform\Form\Form::EVENT_BEFORE_VALIDATE, function (\Solspace\Freeform\Events\Forms\ValidationEvent $e){
+      $form = $e->getForm();
+      if ($form->isValid()){
+        $params = [
+          'content' => []
+        ];
+        foreach($form->getFields() as $field){
+          switch(get_class($field)){
+            case 'Solspace\Freeform\Fields\Implementations\EmailField':
+              $params['email'] = $field->getValue();
+              break;
+            case 'Solspace\Freeform\Fields\Implementations\TextareaField':
+              $params['content'][] = $field->getValue();
+              break;
+          }
         }
-      }
-      if ((OOPSpam::$plugin->settings->enableContextual) && (!empty(OOPSpam::$plugin->settings->contextualContent)) && (in_array($this->integration, OOPSpam::$plugin->settings->contextual))){
-        $params['contextual'] = true;
-      }
-      if (!OOPSpam::$plugin->antiSpam->checkSpam($params, $this->getName())){
-        $e->getForm()->markAsSpam('OOPSpam', 'Blocked by OOPSpam');
+        if ((OOPSpam::$plugin->settings->enableContextual) && (!empty(OOPSpam::$plugin->settings->contextualContent)) && (in_array($this->integration, OOPSpam::$plugin->settings->contextual))){
+          $params['contextual'] = true;
+        }
+        if (!OOPSpam::$plugin->antiSpam->checkSpam($params, $this->getName())){
+          $form->markAsSpam('OOPSpam', 'Blocked by OOPSpam');
+        }
       }
     });
   }